chore(icons): move src/loaders to src
