+++ packages/icons/src/BaseLoader.ts
